Amend service recipients (#1023)

+++ tests/NHSD.GPIT.BuyingCatalogue.E2E.PublicBrowseTests/Areas/Ordering/SolutionSelection/AdditionalServices/SelectAdditionalServiceRecipientsAmendment.cs
@@ -12,14 +12,14 @@ using NHSD.GPIT.BuyingCatalogue.WebApp.Areas.Orders.Controllers.SolutionSelectio
 using NHSD.GPIT.BuyingCatalogue.WebApp.Areas.Orders.Validators.SolutionSelection.ServiceRecipients;
 using Xunit;
 
-namespace NHSD.GPIT.BuyingCatalogue.E2ETests.Areas.Ordering.SolutionSelection.AssociatedServices
+namespace NHSD.GPIT.BuyingCatalogue.E2ETests.Areas.Ordering.SolutionSelection.AdditionalServices
 {
-    public class SelectAssociatedServiceRecipients : BuyerTestBase, IClassFixture<LocalWebApplicationFactory>, IDisposable
+    public class SelectAdditionalServiceRecipientsAmendment : BuyerTestBase, IClassFixture<LocalWebApplicationFactory>, IDisposable
     {
         private const string InternalOrgId = "CG-03F";
-        private const int OrderId = 90008;
-        private static readonly CallOffId CallOffId = new(OrderId, 1);
-        private static readonly CatalogueItemId CatalogueItemId = new(99998, "S-999");
+        private const int OrderNumber = 90031;
+        private static readonly CallOffId CallOffId = new(OrderNumber, 2);
+        private static readonly CatalogueItemId CatalogueItemId = new(99998, "001A99");
 
         private static readonly Dictionary<string, string> Parameters = new()
         {
@@ -28,15 +28,15 @@ namespace NHSD.GPIT.BuyingCatalogue.E2ETests.Areas.Ordering.SolutionSelection.As
             { nameof(CatalogueItemId), $"{CatalogueItemId}" },
         };
 
-        public SelectAssociatedServiceRecipients(LocalWebApplicationFactory factory)
+        public SelectAdditionalServiceRecipientsAmendment(LocalWebApplicationFactory factory)
             : base(factory, typeof(ServiceRecipientsController), nameof(ServiceRecipientsController.AddServiceRecipients), Parameters)
         {
         }
 
         [Fact]
-        public void SelectAssociatedServiceRecipients_AllSectionsDisplayed()
+        public void SelectAdditionalServiceRecipientsAmendment_AllSectionsDisplayed()
         {
-            CommonActions.PageTitle().Should().BeEquivalentTo("Service Recipients for Associated Service - E2E Single Price Added Associated Service".FormatForComparison());
+            CommonActions.PageTitle().Should().BeEquivalentTo("Service Recipients for Additional Service - E2E Multiple Prices Additional Service".FormatForComparison());
             CommonActions.GoBackLinkDisplayed().Should().BeTrue();
             CommonActions.ElementIsDisplayed(ServiceRecipientObjects.PreSelectedInset).Should().BeTrue();
             CommonActions.ElementIsDisplayed(ServiceRecipientObjects.SelectAllLink).Should().BeTrue();
@@ -45,7 +45,7 @@ namespace NHSD.GPIT.BuyingCatalogue.E2ETests.Areas.Ordering.SolutionSelection.As
         }
 
         [Fact]
-        public void SelectAssociatedServiceRecipients_ClickGoBackLink_ExpectedResult()
+        public void SelectAdditionalServiceRecipientsAmendment_ClickGoBackLink_ExpectedResult()
         {
             CommonActions.ClickGoBackLink();
 
@@ -55,7 +55,7 @@ namespace NHSD.GPIT.BuyingCatalogue.E2ETests.Areas.Ordering.SolutionSelection.As
         }
 
         [Fact]
-        public void SelectAssociatedServiceRecipients_ClickSelectAllLink_ExpectedResult()
+        public void SelectAdditionalServiceRecipientsAmendment_ClickSelectAllLink_ExpectedResult()
         {
             CommonActions.ClickLinkElement(ServiceRecipientObjects.SelectAllLink);
 
@@ -77,7 +77,7 @@ namespace NHSD.GPIT.BuyingCatalogue.E2ETests.Areas.Ordering.SolutionSelection.As
         }
 
         [Fact]
-        public void SelectAssociatedServiceRecipients_NoSelectionMade_DisplaysError()
+        public void SelectAdditionalServiceRecipientsAmendment_NoSelectionMade_DisplaysError()
         {
             CommonActions.ClickSave();
 
@@ -94,7 +94,7 @@ namespace NHSD.GPIT.BuyingCatalogue.E2ETests.Areas.Ordering.SolutionSelection.As
         }
 
         [Fact]
-        public void SelectAssociatedServiceRecipients_SelectionMade_ExpectedResult()
+        public void SelectAdditionalServiceRecipientsAmendment_SelectionMade_ExpectedResult()
         {
             GetOrderItem().OrderItemRecipients.Count.Should().Be(0);
 
@@ -111,8 +111,14 @@ namespace NHSD.GPIT.BuyingCatalogue.E2ETests.Areas.Ordering.SolutionSelection.As
         public void Dispose()
         {
             var context = GetEndToEndDbContext();
+            var orderItem = GetOrderItem();
 
-            foreach (var recipient in GetOrderItem().OrderItemRecipients)
+            if (orderItem == null)
+            {
+                return;
+            }
+
+            foreach (var recipient in orderItem.OrderItemRecipients)
             {
                 context.OrderItemRecipients.Remove(recipient);
             }
@@ -123,10 +129,11 @@ namespace NHSD.GPIT.BuyingCatalogue.E2ETests.Areas.Ordering.SolutionSelection.As
         private OrderItem GetOrderItem()
         {
             var context = GetEndToEndDbContext();
+            var orderId = context.OrderId(InternalOrgId, CallOffId).Result;
 
             return context.OrderItems
                 .Include(x => x.OrderItemRecipients)
-                .First(x => x.OrderId == OrderId
+                .FirstOrDefault(x => x.OrderId == orderId
                     && x.CatalogueItemId == CatalogueItemId);
         }
     }
